Introduce `TransactionStep` type and `TransactionSteps` type in `_internal/types.ts` to manage transaction states.
